backend/requirements.txt
# blinker==1.5
# Brlapi==0.8.4
# certifi==2022.9.24
# chardet==5.1.0
# charset-normalizer==3.0.1
# cryptography==38.0.4
# cupshelpers==1.0
# dbus-python==1.3.2
# distro==1.8.0
# distro-info==1.5+deb12u1
# httplib2==0.20.4
# idna==3.3
# lazr.restfulclient==0.14.5
# lazr.uri==1.0.6
# louis==3.24.0
# Mako==1.2.4.dev0
# MarkupSafe==2.1.2
# oauthlib==3.2.2
# pycairo==1.20.1
# pycups==2.0.1
# pycurl==7.45.2
# PyGObject==3.42.2
# PyJWT==2.6.0
# pyparsing==3.0.9
# PySimpleSOAP==1.16.2
# pysmbc==1.0.23
# python-apt==2.6.0
# python-dateutil==2.8.2
# python-debian==0.1.49
# python-debianbts==4.0.1
# pyxdg==0.28
# reportbug==12.0.0
# requests==2.28.1
# six==1.16.0
# ufw==0.36.2
# urllib3==1.26.12
# wadllib==1.3.6
# xdg==5


# Django dependencies
Django>=5.1.4
djangorestframework==3.15.2
channels==4.2.0
django-csp==3.8
daphne>=4.1.2
channels_redis
